Calculator.py
The user will be able to enter numbers and select the desired arithmetic operations. The program will loop until the user chooses to exit the program. Be aware that this is a textual program that works like a calculator, but it is not a graphical program that in any way looks like a calculator.