Montecarlo.py
Monte Carlo is a statistical method using random sampling that is used to perform an estimation of something that is difficult or impossible to measure. This process was named after the city of Monte Carlo in Monaco, which is known for its casinos (and hence randomness, or more properly unpredictability). I will use the Monte Carlo sampling process to estimate the area of circle.