Nbody.py
This project is a discrete simulation solution of an N-body gravitational force problem. Given some number of ‘bodies’ in empty space, the bodies are allowed to move freely and interact with each other under gravitational force. This simulation does not simulate collisions between the bodies. The results are displayed using an animation.