Errorcorrection.py
This project uses number arrays to create groups of bits (binary digits) that are sent over a simulated communications channel from one computer to another. The sender uses parity bits to encode extra information in the data that allows the receiver to detect and correct errors in the transmitted bits. The simulated communications channel will be a function that, with low probability, flips bits randomly.